+++ fev/fev_tmsvc_modules.h
@@ -10,12 +10,17 @@
 #endif
 
 #include <time.h>
+#include <stdint.h>
 
 #include "flibs/fev.h"
 #include "flibs/fev_tmsvc_types.h"
 
+#define NS_PER_SECOND 1000000000L
+#define NS_PER_MS     1000000L
+#define MS_PER_SECOND 1000
+
 #define TIMEMS(tm) \
-    (tm.tv_sec * MS_PER_SECOND + tm.tv_nsec / NS_PER_MS)
+    ((tm).tv_sec * MS_PER_SECOND + (tm).tv_nsec / NS_PER_MS)
 
 // The struct used for contain the basic information from user input
 struct _ftimer_node {
@@ -23,38 +28,38 @@ struct _ftimer_node {
     ftimer_cb       cb;
     void*           arg;
     fev_timer_svc*  owner;
+
+    long            interval;
     long            expiration;
-    int             isvalid;
-    int             _padding;
+
+    uint32_t        valid     :1;
+    uint32_t        triggered :1;
+
+    uint32_t        _padding  :30;
+    int             _padding1;
 };
 
 // This operation table control the low level data storage/operation behaivor
 // It's used for the library maintainers who can add/modify the new/exist modules
-typedef struct fev_timer_svc_opt {
+typedef struct fev_timer_svc_ops_t {
     int  (*init)   (void** mod_data);
     void (*destroy)(void*  mod_data);
     int  (*process)(fev_state*,   void* mod_data, struct timespec* now);
     int  (*add)    (ftimer_node*, void* mod_data);
     int  (*del)    (ftimer_node*, void* mod_data);
-    ftimer_node* (*first)(void* mod_data);
-} fev_tmsvc_opt;
+    ftimer_node* (*top)(void* mod_data);
+} fev_tmsvc_ops_t;
 
 /**
  * @desc If 'now' >= 'start + expire', means timeout
- *
- * @param start: the time of begin
- * @param now: the time of now
- * @param expire: the expiration time, unit is million second
  *
  * @return
  *  - 1: means already timeout
  *  - 0: means doesn't timeout
  */
-int fev_tmmod_timeout(struct timespec* start,
-                      struct timespec* now,
-                      long expiration);
+int fev_tmmod_timeout(ftimer_node* node, struct timespec* now);
 
-extern fev_tmsvc_opt* tmsvc_opt_tbl[];
+extern fev_tmsvc_ops_t* tmsvc_ops_tbl[];
 
 #endif